Split App into subcomponents
